chore: rename filename of `ServiceWorkerCompilerPlugin`
