+++ packages/devtools/client/plugins/ServiceWorkerCompilerPlugin.ts
